+++ crates/zng-app/src/render.rs
@@ -1790,7 +1790,7 @@ impl FrameBuilder {
 
     /// Create a new display list builder that can be built in parallel and merged back onto this one using [`parallel_fold`].
     ///
-    /// Note that split list must be folded before any open reference frames, stacking contexts or clips are closed in this list.
+    /// Note that split list must be folded before any current open reference frames, stacking contexts or clips are closed in this list.
     ///
     /// [`parallel_fold`]: Self::parallel_fold
     pub fn parallel_split(&self) -> ParallelBuilder<Self> {
@@ -1845,6 +1845,73 @@ impl FrameBuilder {
 
         self.widget_count += split.widget_count;
         self.debug_dot_overlays.extend(split.debug_dot_overlays);
+    }
+
+    /// Calls `render` to render a separate nested window on this frame.
+    #[allow(clippy::too_many_arguments)]
+    pub fn with_nested_window(
+        &mut self,
+        render_widgets: Arc<RenderUpdates>,
+        render_update_widgets: Arc<RenderUpdates>,
+
+        root_id: WidgetId,
+        root_bounds: &WidgetBoundsInfo,
+        info_tree: &WidgetInfoTree,
+        default_font_aa: FontAntiAliasing,
+
+        render: impl FnOnce(&mut Self),
+    ) {
+        // similar to parallel_split, but without parent context
+        let mut nested = Self::new(
+            render_widgets,
+            render_update_widgets,
+            self.frame_id,
+            root_id,
+            root_bounds,
+            info_tree,
+            self.renderer.clone(),
+            self.scale_factor,
+            default_font_aa,
+        );
+        nested.display_list = self.display_list.parallel_split();
+        nested.hit_clips = self.hit_clips.parallel_split();
+        nested.widget_count_offsets = self.widget_count_offsets.parallel_split();
+
+        render(&mut nested);
+
+        // finalize nested window
+        info_tree.root().bounds_info().set_rendered(
+            Some(WidgetRenderInfo {
+                visible: nested.visible,
+                parent_perspective: nested.perspective,
+                seg_id: 0,
+                back: 0,
+                front: nested.widget_count,
+            }),
+            info_tree,
+        );
+        info_tree.after_render(
+            nested.frame_id,
+            nested.scale_factor,
+            Some(
+                nested
+                    .renderer
+                    .as_ref()
+                    .and_then(|r| r.generation().ok())
+                    .unwrap_or(ViewProcessGen::INVALID),
+            ),
+            Some(nested.widget_count_offsets.clone()),
+        );
+
+        // fold nested window into host window
+        self.hit_clips.parallel_fold(nested.hit_clips);
+        self.display_list.parallel_fold(nested.display_list);
+
+        self.widget_count_offsets
+            .parallel_fold(nested.widget_count_offsets, self.widget_count);
+
+        self.widget_count += nested.widget_count;
+        self.debug_dot_overlays.extend(nested.debug_dot_overlays);
     }
 
     /// Finalizes the build.
@@ -2632,6 +2699,39 @@ impl FrameUpdate {
         if let Some(c) = self.clear_color.take() {
             self.clear_color = Some(c);
         }
+    }
+
+    /// Calls `update` to render update a separate nested window on this frame.
+    pub fn with_nested_window(
+        &mut self,
+        render_update_widgets: Arc<RenderUpdates>,
+        root_id: WidgetId,
+        root_bounds: WidgetBoundsInfo,
+        update: impl FnOnce(&mut Self),
+    ) {
+        let mut nested = Self::new(
+            render_update_widgets,
+            self.frame_id,
+            root_id,
+            root_bounds,
+            None,
+            self.current_clear_color,
+        );
+
+        update(&mut nested);
+
+        // fold
+        fn take_or_append<T>(t: &mut Vec<T>, s: &mut Vec<T>) {
+            if t.is_empty() {
+                *t = mem::take(s);
+            } else {
+                t.append(s)
+            }
+        }
+        take_or_append(&mut self.transforms, &mut nested.transforms);
+        take_or_append(&mut self.floats, &mut nested.floats);
+        take_or_append(&mut self.colors, &mut nested.colors);
+        take_or_append(&mut self.extensions, &mut nested.extensions);
     }
 
     /// Finalize the update.