+++ crates/zng-ext-window/src/vars.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use zng_app::{
-    widget::info::access::AccessEnabled,
+    widget::{info::access::AccessEnabled, WidgetId},
     window::{MonitorId, WindowId, WINDOW},
 };
 use zng_color::LightDark;
@@ -65,6 +65,7 @@ pub(super) struct WindowVarsData {
     taskbar_visible: ArcVar<bool>,
 
     parent: ArcVar<Option<WindowId>>,
+    pub(super) nest_parent: ArcVar<Option<WidgetId>>,
     modal: ArcVar<bool>,
     pub(super) children: ArcVar<IdSet<WindowId>>,
 
@@ -109,7 +110,7 @@ impl WindowVars {
             position: var(Point::default()),
             monitor: var(MonitorQuery::default()),
             video_mode: var(VideoMode::default()),
-            size: var(Size::new(800, 600)),
+            size: var(Size::default()),
 
             font_size: var(11.pt()),
 
@@ -144,6 +145,7 @@ impl WindowVars {
             taskbar_visible: var(true),
 
             parent: var(None),
+            nest_parent: var(None),
             modal: var(false),
             children: var(IdSet::default()),
 
@@ -582,6 +584,15 @@ impl WindowVars {
     /// [`accent_color`]: Self::accent_color
     pub fn parent(&self) -> ArcVar<Option<WindowId>> {
         self.0.parent.clone()
+    }
+
+    /// Gets the widget in [`parent`] that hosts the window, if it is nesting.
+    ///
+    /// Nesting windows are presented as an widget, similar to an "iframe".
+    ///
+    /// [`parent`]: Self::parent
+    pub fn nest_parent(&self) -> ReadOnlyArcVar<Option<WidgetId>> {
+        self.0.nest_parent.read_only()
     }
 
     /// Defines the [`parent`](Self::parent) connection.